refactor event results

+++ server/api/event-criteria/[id].delete.js
@@ -5,6 +5,16 @@ export default defineEventHandler(async (event) => {
 	const { id } = event.context.params;
 
 	try {
+		const { data: criteriaExist } = await client
+			.from("event_criteria")
+			.select("event_id")
+			.eq("id", id)
+			.maybeSingle();
+
+		if (!criteriaExist) {
+			throw new Error("Criteria not found.");
+		}
+
 		const { data: criteriaData, error: criteriaError } = await client
 			.from("event_criteria")
 			.delete()
@@ -13,6 +23,11 @@ export default defineEventHandler(async (event) => {
 		if (criteriaError) {
 			throw new Error(criteriaError.message);
 		}
+
+		const { error: DeleteError } = await client
+			.from("event_results")
+			.delete()
+			.eq("event_id", criteriaExist.event_id);
 
 		return {
 			success: true,
